+++ lua/mini/extra.lua
@@ -384,6 +384,73 @@ MiniExtra.pickers.buf_lines = function(local_opts, opts)
   return H.pick_start(items, { source = default_source }, opts)
 end
 
+--- Color scheme picker
+---
+--- Pick and apply color scheme. Preview temporarily applies item's color scheme
+--- and shows how selected highlight groups look.
+--- Canceling reverts to color scheme before picker start:
+--- - With |MiniColors-colorscheme:apply()| if |mini.colors| was available.
+--- - With |:colorscheme| if |g:colors_name| was available.
+---
+---@param local_opts __extra_pickers_local_opts
+---   Possible fields:
+---   - <names> `(table)` - array of color scheme names to pick from.
+---     Default: all available color schemes.
+---   - <preview_hl_groups> `(table)` - array of highlight groups to show in preview
+---     window. Default: all defined highlight groups in alphabetical order.
+---
+---@param opts __extra_pickers_opts
+---
+---@return __extra_pickers_return
+MiniExtra.pickers.colorschemes = function(local_opts, opts)
+  local pick = H.validate_pick('colorschemes')
+  local_opts = local_opts or {}
+
+  -- Infer data to show
+  local all_cs = vim.fn.getcompletion('', 'color')
+  local items = local_opts.names or all_cs
+  if not H.islist(items) then H.error('`names` should be array of color scheme names') end
+  for _, item in ipairs(items) do
+    if not vim.tbl_contains(all_cs, item) then H.error(vim.inspect(item) .. ' is not a color scheme name') end
+  end
+
+  local hl_groups = local_opts.preview_hl_groups
+  if hl_groups ~= nil and not H.islist(hl_groups) then H.error('`preview_hl_groups` should be array') end
+
+  -- Compute original color scheme to restore
+  local bg_orig = vim.o.background
+  local has_minicolors, minicolors = pcall(require, 'mini.colors')
+  local cs_orig = has_minicolors and minicolors.get_colorscheme() or vim.g.colors_name
+  local restore = function()
+    vim.o.background = bg_orig
+    if cs_orig == nil then return end
+    if type(cs_orig) == 'string' then return vim.cmd('colorscheme ' .. cs_orig) end
+    cs_orig:apply()
+    -- Trigger to indicate actual color scheme application
+    vim.api.nvim_exec_autocmds('ColorScheme', {})
+  end
+
+  -- Define source
+  local choose = function(item) vim.cmd('colorscheme ' .. item) end
+  local choose_marked = function(items_to_mark) choose(items_to_mark[1] or '') end
+
+  local needs_restore = false
+  local preview = function(buf_id, item)
+    needs_restore = true
+    -- Ensure that previewed color scheme tries to use the initial background.
+    -- This can be not the case if previous preview forced new background.
+    vim.o.background = bg_orig
+    vim.cmd('colorscheme ' .. item)
+    H.preview_cs_hl_groups(buf_id, hl_groups)
+  end
+
+  local default_source = { name = 'Colorschemes', preview = preview, choose = choose, choose_marked = choose_marked }
+  local result = H.pick_start(items, { source = default_source }, opts)
+
+  if result == nil and needs_restore then restore() end
+  return result
+end
+
 --- Neovim commands picker
 ---
 --- Pick from Neovim built-in (|ex-commands|) and |user-commands|.
@@ -1712,6 +1779,22 @@ H.choose_with_buflisted = function(item)
   local win_target = pick.get_picker_state().windows.target
   local buf_id = vim.api.nvim_win_get_buf(win_target)
   vim.bo[buf_id].buflisted = true
+end
+
+-- Colorscheme picker ----------------------------------------------------------
+H.preview_cs_hl_groups = function(buf_id, hl_groups)
+  local lines = hl_groups
+  if lines == nil then
+    lines = vim.tbl_keys(vim.api.nvim_get_hl(0, {}))
+    table.sort(lines)
+  end
+  vim.api.nvim_buf_set_lines(buf_id, 0, -1, false, lines)
+
+  local ns_id = H.ns_id.pickers
+  vim.api.nvim_buf_clear_namespace(buf_id, ns_id, 0, -1)
+  for i, hl in ipairs(lines) do
+    pcall(vim.api.nvim_buf_set_extmark, buf_id, ns_id, i - 1, 0, { end_row = i, end_col = 0, hl_group = hl })
+  end
 end
 
 -- Diagnostic picker ----------------------------------------------------------