+++ tests/dir-extra/colorschemes/colors/miniforcebg.lua
@@ -1,0 +1,2 @@
+vim.o.background = vim.o.background == 'dark' and 'light' or 'dark'
+vim.g.colors_name = 'miniforcebg'